Implementação de testes de api para feature entregadores.
