--- backend/tests/controllers/deliveryman.service.steps.ts
@@ -1,194 +1,0 @@
-import { loadFeature, defineFeature } from 'jest-cucumber';
-import DeliverymanRepository from '../../src/repositories/deliverymans/deliveryman.repository';
-import supertest from 'supertest';
-import app from '../../src/app';
-import { di } from '../../src/di';
-
-const feature = loadFeature('tests/features/deliverymans/Register.feature', {tagFilter: "@read"});
-const request = supertest(app);
-
-defineFeature(feature, (test) => {
-    // mocking the repository
-    let mockDelDeliverymanRepository:DeliverymanRepository;
-    let response: supertest.Response;
-  
-    beforeEach(() => {
-      mockDelDeliverymanRepository = di.getRepository<DeliverymanRepository>(DeliverymanRepository);
-    });
-  
-    test('Obtenção de todos os entregadores', ({ given, when, then, and }) => {
-      given(/^o sistema já possui o entregador com "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)" e "(.*)" preenchido com "(.*)" e o entregador com "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)" e "(.*)" preenchido com "(.*)"$/, 
-      async (id, idValue, name, nameValue, email, emailValue, numOrders, numOrdersValue, numRates, numRatesValue, id2, idValue2, name2, nameValue2, email2, emailValue2, numOrders2, numOrdersValue2, numRates2, numRatesValue2) => {
-        // Check if the deliveryman does not exist in the repository and delete it if it exists
-        const existingDeliveryman = await mockDelDeliverymanRepository.getDeliveryman(id);
-        if (!existingDeliveryman) {
-          response = await request.post("/api/entregadores").send({
-            id: idValue,
-            name: nameValue, 
-            email: emailValue
-          });
-        }
-        const existingDeliveryman2 = await mockDelDeliverymanRepository.getDeliveryman(id2);
-        if (!existingDeliveryman2) {
-          response = await request.post("/api/entregadores").send({
-            id: idValue2,
-            name: nameValue2, 
-            email: emailValue2
-          });
-        }
-      });
-  
-      when(
-        /^uma requisição GET for enviada para "(.*)"$/,
-        async (url) => {
-          response = await request.get(url);
-        }
-      );
-  
-      then(/^o status da resposta deve ser "(.*)"$/, (statusCode) => {
-        expect(response.status).toBe(parseInt(statusCode, 10));
-      });
-
-      and(/^o campo "(.*)" do corpo da resposta deve estar preenchido com "(.*)"$/, (msgCode, msgCodeValue) => {
-        expect(response.body.msgCode).toBe(msgCodeValue);
-      });
-  
-      and(/^o JSON da resposta deve conter o entregador com o campo "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)" e "(.*)" preenchido com "(.*)" e o entregador com o campo "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)" e "(.*)" preenchido com "(.*)"$/, 
-        (id, idValue, name, nameValue, email, emailValue, numOrders, numOrdersValue, numRates, numRatesValue, id2, idValue2, name2, nameValue2, email2, emailValue2, numOrders2, numOrdersValue2, numRates2, numRatesValue2) => {
-          expect(response.body.data).toEqual([{"email": "aninha@getMaxListeners.com", "id": "1", "name": "Ana Santana", "numOrders": "0", "numRates": "0"}, {"email": "iza@yahoo.com", "id": "452", "name": "Izabel", "numOrders": "0", "numRates": "0"}, {"email": "maria@gmail.com", "id": "124", "name": "Maria", "numOrders": "0", "numRates": "0"}]);;
-        }
-      );
-    });
-
-    test('Obtenção de um entregador por ID', ({ given, when, then, and }) => {
-      given(/^o sistema já possui um entregador com "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)" e "(.*)" preenchido com "(.*)"$/, 
-      async (id, idValue, name, nameValue, email, emailValue, numOrders, numOrdersValue, numRates, numRatesValue) => {
-        // Check if the deliveryman does not exist in the repository and delete it if it exists
-        const existingDeliveryman = await mockDelDeliverymanRepository.getDeliveryman(id);
-        if (!existingDeliveryman) {
-          response = await request.post("/api/entregadores").send({
-            id: idValue,
-            name: nameValue, 
-            email: emailValue
-          });
-        }
-      });
-  
-      when(
-        /^uma requisição GET for enviada para "(.*)"$/,
-        async (url) => {
-          response = await request.get(url);
-        }
-      );
-  
-      then(/^o status da resposta deve ser "(.*)"$/, (statusCode) => {
-        expect(response.status).toBe(parseInt(statusCode, 10));
-      });
-
-      and(/^o campo "(.*)" do corpo da resposta deve estar preenchido com "(.*)"$/, (msgCode, msgCodeValue) => {
-        expect(response.body.msgCode).toBe(msgCodeValue);
-      });
-  
-      and(/^o JSON da resposta deve conter um entregador com o campo "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)" e "(.*)" preenchido com "(.*)"$/, 
-        (id, idValue, name, nameValue, email, emailValue, numOrders, numOrdersValue, numRates, numRatesValue) => {
-          expect(response.body.data).toEqual(
-            expect.objectContaining({
-                id: idValue,
-                name: nameValue, 
-                email: emailValue,
-                numOrders: numOrdersValue,
-                numRates: numRatesValue
-            })
-          );
-        }
-      );
-    });
-    
-    test('Criação de um entregador bem sucedida', ({ given, when, then, and }) => {
-      given(/^o sistema não tem um entregador com "(.*)" igual a "(.*)"$/, async (id, idValue) => {
-        // Check if the deliveryman does not exist in the repository and delete it if it exists
-        const existingDeliveryman = await mockDelDeliverymanRepository.getDeliveryman(id);
-        if (existingDeliveryman) {
-          await mockDelDeliverymanRepository.deleteDeliveryman;
-        }
-      });
-  
-      when(
-        /^uma requisição POST for enviada para "(.*)" com o corpo da requisição sendo um JSON com o campo "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)" e "(.*)" preenchido com "(.*)"$/,
-        async (url, id, idValue, name, nameValue, email, emailValue) => {
-          response = await request.post(url).send({
-            id: idValue,
-            name: nameValue, 
-            email: emailValue 
-          });
-        }
-      );
-  
-      then(/^o status da resposta deve ser "(.*)"$/, (statusCode) => {
-        expect(response.status).toBe(parseInt(statusCode, 10));
-      });
-
-      and(/^o campo "(.*)" do corpo da resposta deve estar preenchido com "(.*)"$/, (msgCode, msgCodeValue) => {
-        expect(response.body.msgCode).toBe(msgCodeValue);
-      });
-  
-      and(/^o JSON da resposta deve conter o campo "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)" e "(.*)" preenchido com "(.*)"$/, 
-        (id, idValue, name, nameValue, email, emailValue, numOrders, numOrdersValue, numRates, numRatesValue) => {
-          expect(response.body.data).toEqual(
-            expect.objectContaining({
-                id: idValue,
-                name: nameValue, 
-                email: emailValue,
-                numOrders: numOrdersValue,
-                numRates: numRatesValue
-            })
-          );
-        }
-      );
-    });
-
-    test('Criação de um entregador já existente', ({ given, when, then, and }) => {
-      given(/^o sistema já possui um entregador com "(.*)" igual a "(.*)"$/, async (id, idValue) => {
-        // Check if the deliveryman does not exist in the repository and delete it if it exists
-        const existingDeliveryman = await mockDelDeliverymanRepository.getDeliveryman(id);
-        if (!existingDeliveryman) {
-          response = await request.post("/api/entregadores").send({
-            id: idValue,
-            name: "Pessoa", 
-            email: "pessoa@gmail.com" 
-          });
-        }
-      });
-  
-      when(
-        /^uma requisição POST for enviada para "(.*)" com o corpo da requisição sendo um JSON com o campo "(.*)" preenchido com "(.*)", "(.*)" preenchido com "(.*)" e "(.*)" preenchido com "(.*)"$/,
-        async (url, id, idValue, name, nameValue, email, emailValue) => {
-          response = await request.post(url).send({
-            id: idValue,
-            name: nameValue, 
-            email: emailValue 
-          });
-        }
-      );
-  
-      then(/^o status da resposta deve ser "(.*)"$/, (statusCode) => {
-        expect(response.status).toBe(parseInt(statusCode, 10));
-      });
-
-      and(/^o campo "(.*)" do corpo da resposta deve estar preenchido com "(.*)"$/, (msgCode, msgCodeValue) => {
-        expect(response.body.msgCode).toBe(msgCodeValue);
-      });
-  
-      and(/^o JSON da resposta deve ser nulo$/, 
-        (id, idValue, name, nameValue, email, emailValue, numOrders, numOrdersValue, numRates, numRatesValue) => {
-          expect(response.body.data).toEqual(
-            expect.objectContaining({})
-          );
-        }
-      );
-    });
-
-    
-
-    
-  });
